Create snake_game.c
Snake Game is a simple console application without graphics. Foods are provided at the several co-ordinates of the screen for the snake to eat. Every time the snake eats the food, its length will by increased by one element along with the score. Snake game gives users a total of three lives to play the game. The life-count decreases as the snake hits the wall or its own body.